Rework how we deal with ssh keys for containers.
Make it so that images are always saved with the keys given in the image
spec. Then we always use those same keys to connect to a new container
from that spec, and when creating the container we then log in and
update the admin user setup to be compatible with the host-config.